+++ lib/module_implementations/definition_impl_devel.c
@@ -825,6 +825,7 @@ static const struct def_algo_prereqs devel_kdf_tpm_prereqs[] = {
 /**************************************************************************
  * TLS Definitions
  **************************************************************************/
+#if 0
 static const struct def_algo_prereqs devell_kdf_prereqs[] = {
 	{
 		.algorithm = "HMAC",
@@ -859,6 +860,30 @@ static const struct def_algo_prereqs devell_kdf_prereqs[] = {
 	DEVEL_KDF11,							\
 	DEVEL_KDF12
 
+#else
+#define DEVEL_KDF
+#endif
+
+/**************************************************************************
+ * ANSI X963 Definitions
+ **************************************************************************/
+#if 1
+#define DEVEL_ANSI_X963							\
+	{								\
+	.type = DEF_ALG_TYPE_ANSI_X963,					\
+	.algo.ansi_x963 = {						\
+		.hashalg = ACVP_SHA256 | ACVP_SHA384,			\
+		DEF_ALG_DOMAIN(.shared_info_len, 0, 1024, 8),		\
+		DEF_ALG_DOMAIN(.key_data_len, 128, 4096, 8),		\
+		.field_size[0] = 256,					\
+		.field_size[1] = 283,					\
+		}							\
+	}
+
+#else
+#define DEVEL_ANSI_X963
+#endif
+
 /**************************************************************************
  * Devel Implementation Definitions
  **************************************************************************/
@@ -892,6 +917,8 @@ static const struct def_algo devel[] = {
 	DEVEL_COND_COMP
 
 	DEVEL_KDF
+
+	DEVEL_ANSI_X963
 };
 
 /**************************************************************************