+++ tests/requests/definition_impl_test_ansi_x963.c
@@ -1,0 +1,63 @@
+/*
+ * Copyright (C) 2021, Stephan Mueller <smueller@chronox.de>
+ *
+ * License: see LICENSE file in root directory
+ *
+ * THIS SOFTWARE IS PROVIDED ``AS IS'' AND ANY EXPRESS OR IMPLIED
+ * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE, ALL OF
+ * WHICH ARE HEREBY DISCLAIMED.  IN NO EVENT SHALL THE AUTHOR BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT
+ * OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
+ * BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+ * USE OF THIS SOFTWARE, EVEN IF NOT ADVISED OF THE POSSIBILITY OF SUCH
+ * DAMAGE.
+ */
+
+#include "definition.h"
+#include "definition_impl_common.h"
+
+/**************************************************************************
+ * ANSI X9.63 Definitions
+ **************************************************************************/
+
+#define TEST_ANSI_X963							\
+	{								\
+	.type = DEF_ALG_TYPE_ANSI_X963,					\
+	.algo.ansi_x963 = {						\
+		.hashalg = ACVP_SHA256 | ACVP_SHA384,			\
+		DEF_ALG_DOMAIN(.shared_info_len, 0, 1024, 8),		\
+		DEF_ALG_DOMAIN(.key_data_len, 128, 4096, 8),		\
+		.field_size[0] = 256,					\
+		.field_size[1] = 283,					\
+		}							\
+	}
+
+/**************************************************************************
+ * Tests Generic Definitions
+ **************************************************************************/
+static const struct def_algo tests[] = {
+	TEST_ANSI_X963,
+};
+
+/**************************************************************************
+ * Register operation
+ **************************************************************************/
+
+static struct def_algo_map tests_algo_map [] = {
+	{
+		SET_IMPLEMENTATION(tests),
+		.algo_name = "Tests",
+		.processor = "",
+		.impl_name = "ANSI X9.63"
+	}
+};
+
+ACVP_DEFINE_CONSTRUCTOR(tests_register)
+static void tests_register(void)
+{
+	acvp_register_algo_map(tests_algo_map, ARRAY_SIZE(tests_algo_map));
+}